gxm: Fix DkVtxAttribState unused bits garbage
DkVtxAttribState is defined as:
```c
typedef struct DkVtxAttribState {
    uint32_t bufferId : 5;
    uint32_t : 1;                // <-------------- Unused bit
    uint32_t isFixed : 1;
    uint32_t offset : 14;
    uint32_t size : 6;
    uint32_t type : 3;
    uint32_t : 1;                // <-------------- Unused bit
    uint32_t isBgra : 1;
} DkVtxAttribState;
```
So there are a couple of "unused" bits, and the compiler was free to leave garbage to them (sometimes 0, sometimes 1).
The Tegra GPU cares about those bits and will crash if they are 1 instead of 0.
```
GPU method error (irq 0x00100000), result 0
[b197:458] = 0x78800000, result 0
```
The top unused bit was being set to 1 instead of 0.
`[b197:458]` is `MAXWELL_B` register `NVB197_SET_VERTEX_ATTRIBUTE_A(0)`.